+++ Portal.Application/ModelDTOs/CurrencyDTO.cs
@@ -8,14 +8,14 @@ namespace Portal.Application.ModelDTOs
 {
     public class CurrencyDTO
     {
-        public short CurrencyNumericCode { get; set; }
+        public int CurrencyNumericCode { get; set; }
         public string Entity { get; set; }
         public string CurrencyType { get; set; }
         public string AlphabeticCode { get; set; }
         public decimal ExchangeRate { get; set; }
         public int UserID { get; set; }
 
-        public CurrencyDTO(short currencyNumericCode, string entity, string currencyType, string alphabeticCode, decimal exchangeRate, int userID)
+        public CurrencyDTO(int currencyNumericCode, string entity, string currencyType, string alphabeticCode, decimal exchangeRate, int userID)
         {
             CurrencyNumericCode = currencyNumericCode;
             Entity = entity;
